refactor: remove lv_ prefix from folder names
