+++ src/libs/gltf/gltf_data/lv_gltf_data_injest.cpp
